tab.(JSON)
Return Field Definitions 
searchtermSearchterm supplied to the API though the use of the q parameter. matchtermIdentical to searchterm and include for historical reasons to maintain backwards compatibility. pageId of the current page that the query has returned. This is a zero based index. nextpageId of the offset of the next page. Is always set to +1 the current page. To determine the end or the results query till no results are returned. previouspageId of the offset of the previous page. If no previous page is avaliable will be set to null. This is a zero based index. totalThe total number of results that match the searchterm in the index. Note that this value is approximate. resultsReturns an array containing the matching documentation results. synopsisCode snippet or example of the matching result. displaynameDisplay ready name of the type of result. nameName of the result. urlURL to the result source. Usually is a deep link directly to the documentation. typeSame as display name but in a format allowing for easier grouping. iconLegacy. Was used to display the favicon for the result. namespaceNamespace of the code result if it exists. descriptionA description of the function method or whatever the result is describing.